# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeMaster API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  demo: "demo",
  realization: "realization",
  concept: "concept",
} as const;

export interface Project {
  id: string;
  name: string;
  category: string;
  industry: string;
  description: string;
  features: string[];
  status: ProjectStatus;
  featured: boolean;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  demoUrl?: string | null;
  technologies?: string[];
  createdAt?: string;
}

export type CreateProjectRequestStatus =
  (typeof CreateProjectRequestStatus)[keyof typeof CreateProjectRequestStatus];

export const CreateProjectRequestStatus = {
  demo: "demo",
  realization: "realization",
  concept: "concept",
} as const;

export interface CreateProjectRequest {
  name: string;
  category: string;
  industry: string;
  description: string;
  features: string[];
  status: CreateProjectRequestStatus;
  featured?: boolean;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  demoUrl?: string | null;
  technologies?: string[];
}

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  new: "new",
  in_progress: "in_progress",
  quoted: "quoted",
  closed_won: "closed_won",
  closed_lost: "closed_lost",
} as const;

export interface Lead {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  projectType: string;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  deadline?: string | null;
  description: string;
  status: LeadStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  userId?: string | null;
  createdAt: string;
}

export interface SubmitLeadRequest {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @minLength 1 */
  projectType: string;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @minLength 10 */
  description: string;
  consentContact: boolean;
  consentPrivacy: boolean;
}

export type UpdateLeadRequestStatus =
  (typeof UpdateLeadRequestStatus)[keyof typeof UpdateLeadRequestStatus];

export const UpdateLeadRequestStatus = {
  new: "new",
  in_progress: "in_progress",
  quoted: "quoted",
  closed_won: "closed_won",
  closed_lost: "closed_lost",
} as const;

export interface UpdateLeadRequest {
  status?: UpdateLeadRequestStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  priority?: string | null;
}

export interface UpdateProfileRequest {
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
}

export type AiChatRequestHistoryItemRole =
  (typeof AiChatRequestHistoryItemRole)[keyof typeof AiChatRequestHistoryItemRole];

export const AiChatRequestHistoryItemRole = {
  user: "user",
  assistant: "assistant",
} as const;

export type AiChatRequestHistoryItem = {
  role: AiChatRequestHistoryItemRole;
  content: string;
};

export interface AiChatRequest {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  message: string;
  history?: AiChatRequestHistoryItem[];
}

export interface AiChatResponse {
  reply: string;
}

export type AdminUserRole = (typeof AdminUserRole)[keyof typeof AdminUserRole];

export const AdminUserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface AdminUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  role: AdminUserRole;
  createdAt: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListProjectsParams = {
  category?: string;
  search?: string;
};

export type ListProjects200 = {
  projects: Project[];
};

export type ToggleFavoriteProject200 = {
  favorited: boolean;
};

export type SubmitLead201 = {
  success: boolean;
  message: string;
};

export type ListLeadsParams = {
  status?: string;
};

export type ListLeads200 = {
  leads: Lead[];
};

export type ListUsers200 = {
  users: AdminUser[];
};

export type UpdateUserRoleBodyRole =
  (typeof UpdateUserRoleBodyRole)[keyof typeof UpdateUserRoleBodyRole];

export const UpdateUserRoleBodyRole = {
  user: "user",
  admin: "admin",
} as const;

export type UpdateUserRoleBody = {
  role: UpdateUserRoleBodyRole;
};

export type GetUserFavorites200 = {
  projects: Project[];
};

export type GetUserInquiries200 = {
  inquiries: Lead[];
};