# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeMaster API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    post:
      operationId: submitLead
      tags: [leads]
      summary: Submit a contact/lead form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitLeadRequest"
      responses:
        "201":
          description: Lead submitted and email sent
          content:
            application/json:
              schema:
                type: object
                required: [success, message]
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      operationId: listLeads
      tags: [leads, admin]
      summary: List all leads (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: object
                required: [leads]
                properties:
                  leads:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lead"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /leads/{id}:
    put:
      operationId: updateLead
      tags: [leads, admin]
      summary: Update lead status/notes (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadRequest"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send a message to the AI agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin, users]
      summary: List all users (admin only)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: object
                required: [users]
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/users/{id}/role:
    put:
      operationId: updateUserRole
      tags: [admin, users]
      summary: Update user role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [user, admin]
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Lead:
      type: object
      required: [id, name, email, projectType, description, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: [string, 'null']
        company:
          type: [string, 'null']
        projectType:
          type: string
        budget:
          type: [string, 'null']
        deadline:
          type: [string, 'null']
        description:
          type: string
        status:
          type: string
          enum: [new, in_progress, quoted, closed_won, closed_lost]
        notes:
          type: [string, 'null']
        priority:
          type: [string, 'null']
        userId:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    SubmitLeadRequest:
      type: object
      required: [name, email, projectType, description, consentContact, consentPrivacy]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: [string, 'null']
        company:
          type: [string, 'null']
        projectType:
          type: string
          minLength: 1
        budget:
          type: [string, 'null']
        deadline:
          type: [string, 'null']
        description:
          type: string
          minLength: 10
        consentContact:
          type: boolean
        consentPrivacy:
          type: boolean
    UpdateLeadRequest:
      type: object
      properties:
        status:
          type: string
          enum: [new, in_progress, quoted, closed_won, closed_lost]
        notes:
          type: [string, 'null']
        priority:
          type: [string, 'null']
...
    AiChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 2000
        history:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
    AiChatResponse:
      type: object
      required: [reply]
      properties:
        reply:
          type: string
    AdminUser:
      type: object
      required: [id, email, firstName, lastName, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time