# lib/api-zod/src/generated/types/submitLeadRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CodeMaster API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitLeadRequest {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @minLength 1 */
  projectType: string;
  /** @nullable */
  budget?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @minLength 10 */
  description: string;
  consentContact: boolean;
  consentPrivacy: boolean;
}